a program that translates assembly language into machine language
this program will take a file with assembly code and translate it into
machine code.
the output will a text file that will represent each command and memory
words if needed by base 32

program will also create a file for external labels
and a file with entries if everything is correct on the assembly code.

in case of a syntax , compilation or running error on the assembly code
- the program will create an error log , named exactly as the input file
and will report the found issues.